refactored folder structures, created separate types file per component, moved ScrollToTop into its own component directory, cleaned up Admin components and structure
