+++ src/components/About/index.ts
@@ -1,0 +1,1 @@
+export {About as default} from './About.component'
