+++ src/components/Admin/AdminHome/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AdminHome.component'
