+++ src/components/Admin/AdminLogin/AdminLogin.component.tsx
@@ -5,25 +5,15 @@ import React, {
     useEffect,
     useState,
 } from 'react';
-import { ThunkDispatch } from 'redux-thunk';
-import { backendLogin, backendVerifySession } from 'redux/Admin/Admin.reducer';
 import { connect } from 'react-redux';
-import { RootState } from 'redux/store';
 import { Redirect } from 'react-router';
-import { setJwt } from 'redux/Admin/Admin.action';
+import {
+    AdminLoginPropsAndActions,
+    mapDispatchToProps,
+    mapStateToProps,
+} from './AdminLogin.types';
 
-interface AdminLoginProps {
-    jwt: string;
-}
-
-interface AdminLoginActions {
-    login: (username: string, password: string) => void;
-    verifyAndSetJwt: (inputJwt: string) => void;
-}
-
-type AdminLoginPropsAndActions = AdminLoginProps & AdminLoginActions;
-
-export const AdminLoginComponent: FC<AdminLoginPropsAndActions> = (props) => {
+export const AdminLogin: FC<AdminLoginPropsAndActions> = (props) => {
     const { jwt, login, verifyAndSetJwt } = props;
 
     const [inputUsername, setInputUsername] = useState('');
@@ -84,31 +74,9 @@ export const AdminLoginComponent: FC<AdminLoginPropsAndActions> = (props) => {
     );
 };
 
-export const mapStateToProps = (state: RootState): AdminLoginProps => {
-    return {
-        jwt: state.adminState.jwt,
-    };
-};
 
-export const mapDispatchToProps = (
-    dispatch: ThunkDispatch<{}, {}, any>
-): AdminLoginActions => {
-    return {
-        login: async (username: string, password: string) => {
-            await dispatch(backendLogin(username, password));
-        },
-        verifyAndSetJwt: async (inputJwt: string) => {
-            const verified: any = await dispatch(
-                backendVerifySession(inputJwt)
-            );
-            if (verified) {
-                dispatch(setJwt(inputJwt));
-            }
-        },
-    };
-};
 
-export const AdminLogin = connect(
+export default connect(
     mapStateToProps,
     mapDispatchToProps
-)(AdminLoginComponent);
+)(AdminLogin);
