+++ src/components/Admin/AdminPostCreate/index.ts
@@ -1,0 +1,1 @@
+export {default} from './AdminPostCreate.component'
