+++ src/components/Admin/AdminPostEdit/index.ts
@@ -1,0 +1,1 @@
+export {default} from './AdminPostEdit.component'
