+++ src/components/Admin/index.ts
@@ -1,0 +1,5 @@
+export {default as AdminHome} from './AdminHome';
+export {default as AdminLessons} from './AdminLessons';
+export {default as AdminLogin} from './AdminLogin';
+export {default as AdminPostCreate} from './AdminPostCreate'
+export {default as AdminPostEdit} from './AdminPostEdit'
