+++ src/components/FAQ/index.ts
@@ -1,0 +1,1 @@
+export {FAQ} from './FAQ.component'
