+++ src/components/Lessons/Lessons.component.tsx
@@ -1,59 +1,9 @@
 import React, { FC, Fragment, useCallback, useEffect } from 'react';
-import { RootState } from 'redux/store';
-import {
-    backendGetLessons,
-    backendGetPostsByLessons,
-    Post,
-} from 'redux/Posts/Posts.reducer';
-import { ThunkDispatch } from 'redux-thunk';
 import { connect } from 'react-redux';
-import { RouteComponentProps } from 'react-router';
 import { PostCard } from 'components/PostCard/PostCard.component';
-import { clearPosts, setPostLoading } from 'redux/Posts/Posts.action';
+import {LessonsPropsAndActions, mapDispatchToProps, mapStateToProps} from './Lessons.types';
 
-export interface LessonsProps {
-    posts: Post[];
-    lessons: { id: number; lesson: string }[];
-    postsLoading: boolean;
-}
-
-export interface LessonsActions {
-    getPostsForLessons: (lessons: string[]) => void;
-    clearPosts: () => void;
-    getLessons: () => any;
-    setPostLoading: (loading: boolean) => void;
-}
-
-export type LessonsPropsAndActions = LessonsProps &
-    LessonsActions &
-    RouteComponentProps;
-
-export const mapStateToProps = (state: RootState): LessonsProps => ({
-    posts: state.postState.posts,
-    lessons: state.postState.lessons,
-    postsLoading: state.postState.loadingPosts,
-});
-
-export const mapDispatchToProps = (
-    dispatch: ThunkDispatch<{}, {}, any>
-): LessonsActions => {
-    return {
-        getPostsForLessons: async (lessons: string[]) => {
-            let output = [];
-            for (const l of lessons) {
-                output.push(await dispatch(backendGetPostsByLessons([l])));
-            }
-            return output;
-        },
-        clearPosts: () => dispatch(clearPosts()),
-        getLessons: async () => {
-            return await dispatch(backendGetLessons());
-        },
-        setPostLoading: (loading: boolean) => dispatch(setPostLoading(loading)),
-    };
-};
-
-const LessonsComponent: FC<LessonsPropsAndActions> = (props) => {
+const Lessons: FC<LessonsPropsAndActions> = (props) => {
     const {
         clearPosts,
         posts,
@@ -105,7 +55,7 @@ const LessonsComponent: FC<LessonsPropsAndActions> = (props) => {
     );
 };
 
-export const Lessons = connect(
+export default connect(
     mapStateToProps,
     mapDispatchToProps
-)(LessonsComponent);
+)(Lessons);
