+++ src/components/Lessons/index.ts
@@ -1,0 +1,1 @@
+export {default} from './Lessons.component'
