+++ src/components/PostCard/PostCard.types.ts
@@ -1,0 +1,9 @@
+import {Post} from "../../redux/Posts/Posts.reducer";
+
+export interface PostCardProps {
+    post: Post;
+    onClickTag?: (e: any) => void;
+    onClickCategory?: (e: string) => void;
+    showTitleOnly?: boolean;
+    showPreviewOnly?: boolean;
+}
