+++ src/components/PostCard/index.ts
@@ -1,0 +1,1 @@
+export { PostCard as default } from './PostCard.component';
