+++ src/components/QuizCard/QuizCard.types.ts
@@ -1,0 +1,6 @@
+import {IQuiz} from "../../redux/Posts/Posts.reducer";
+
+export interface QuizCardProps {
+    onCross?: () => void;
+    quiz: IQuiz;
+}
