+++ src/components/QuizCard/index.ts
@@ -1,0 +1,1 @@
+export { QuizCard as default } from './QuizCard.component';
