+++ src/components/Tag/Tag.types.ts
@@ -1,0 +1,4 @@
+export interface TagProps {
+    text: string;
+    onClick?: (e: string) => void;
+}
