+++ src/components/Tag/index.ts
@@ -1,0 +1,1 @@
+export { Tag as default } from './Tag.component';
