+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import { App } from 'components/App/App.component';
+import App from 'components/App';
 import { store } from 'redux/store';
 
 import './index.css';
